refactor(protocol): restructuring protocol Error types and removing unnecessary variants
