+++ safenode/src/network/cmd.rs
@@ -56,7 +56,7 @@ pub enum SwarmCmd {
     /// Get data from the kademlia store
     GetData {
         key: RecordKey,
-        sender: oneshot::Sender<QueryResponse>,
+        sender: oneshot::Sender<Result<QueryResponse>>,
     },
 }
